Moved Dice calaculation to modules
